Added comments and fixed potential empty dir bug

+++ main.py
@@ -7,6 +7,10 @@ from contextlib import contextmanager,redirect_stderr,redirect_stdout
 from os import devnull
 from progress.bar import Bar
 
+# Clear log file
+open('out.log', 'w').close()
+
+# Suppress warning messages
 @contextmanager
 def suppress_stdout_stderr():
     """A context manager that redirects stdout and stderr to devnull"""
@@ -14,6 +18,7 @@ def suppress_stdout_stderr():
         with redirect_stderr(fnull) as err, redirect_stdout(fnull) as out:
             yield (err, out)
 
+# Returns true if song made it to images
 def helper(i, images):
     for j in images:
         if i.replace(".mid", "") in j.replace(".png", ""):
@@ -24,23 +29,28 @@ def helper(i, images):
 files = os.listdir("midiFiles")
 images = os.listdir("imgOut")
 
-with Bar('Cleaning directories', max=len(images)) as bar:
-    for f in images:
-        os.remove(f"imgOut/{f}")
-        bar.next()
+# Cleans up the image directory
+if len(images) > 0:
+    with Bar('Cleaning directories', max=len(images)) as bar:
+        for f in images:
+            os.remove(f"imgOut/{f}")
+            bar.next()
 
 print("\033[032m✓\033[0m Done\n")
 
+# Convert midis to images
 bar = Bar('Converting midi files to images', max=len(files))
 for i in range(len(files)):
      with suppress_stdout_stderr():
         main_midi(f"midiFiles/{files[i]}", 100)
+        open('out.log', 'a').write("\n").close()
         bar.next()
 
 bar.finish()
 
 print("\033[032m✓\033[0m Done\n")
 
+# Convert images to midis
 images = os.listdir("imgOut")
 bar = Bar('Converting filtered images to midi files', max=len(images))
 for i in range(len(images)):
@@ -51,6 +61,7 @@ bar.finish()
 
 print("\033[032m✓\033[0m Done\n\nRemoving all redundant files...")
 
+# Removes midis that did not make it to images to save space
 for i in files:
     if not helper(i, images):
         print(f" ∘ Removing {i}...")
